mamaya

+++ src/main/java/com/roi/cartoh/controller/api/RestController.java
@@ -1,10 +1,12 @@
-package com.roi.cartoh.controller;
+package com.roi.cartoh.controller.api;
 
 import com.roi.cartoh.dto.CarDTO;
 import com.roi.cartoh.model.Car;
 import com.roi.cartoh.service.CarService;
 import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -18,33 +20,41 @@ public class RestController {
         this.carService = carService;
     }
 
-    // GET all cars
+
     @GetMapping("/cars")
     public List<Car> getCars() {
         return carService.findAll();
     }
 
-    // GET car by ID
+
     @GetMapping("/cars/{id}")
     public Car getCarById(@PathVariable Integer id) {
+
         return carService.findById(id);
     }
 
-    // POST new car
+
     @PostMapping("/cars")
     public Car newCar(@Valid @RequestBody CarDTO car) {
         return carService.save(car);
     }
 
-    // PUT update existing car
+
     @PutMapping("/cars/{id}")
-    public Car updateCar(@PathVariable Integer id, @Valid @RequestBody CarDTO car) {
+    public Car updateCar(@PathVariable Integer id, @Valid @RequestBody CarDTO car){
+        Car updateCar = carService.findById(id);
+        if (updateCar == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Car with ID "+ id + " not found.");
+        }
         return carService.updateCar(id, car);
     }
 
-    // DELETE car by ID
+
     @DeleteMapping("/cars/{id}")
     public void deleteCar(@PathVariable Integer id) {
+        if(carService.findById(id) == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Car with ID "+ id + " not found.");
+        }
         carService.deleteCar(id);
     }
 }
